trending-function/requirements.txt
google-cloud-bigquery
google-cloud-firestore
functions-framework
pandas
numpy
scikit-learn
hdbscan
db-dtypes